moved token generation code for Create Profile to reduce redundancy.
